undo changes

+++ pw-bordercontrol/cmd/bordercontrol/containers.go
@@ -44,15 +44,14 @@ func checkFeederContainers(ctx *cli.Context) {
 	}
 	defer cli.Close()
 
-	// prepare filter to find feed-in containers
+	// prepare filters to find feed-in containers
 	cfcLog.Info().Msg("prepare filter to find feed-in containers")
-	filters := filters.NewArgs()
-	filters.Add("name", "feed-in-*")
-	filters.Add("label", fmt.Sprintf("image!=%s", ctx.String("feedinimage")))
+	filterFeedIn := filters.NewArgs()
+	filterFeedIn.Add("name", "feed-in-*")
 
 	// find containers
 	cfcLog.Info().Msg("find containers")
-	containers, err := cli.ContainerList(dockerCtx, types.ContainerListOptions{Filters: filters})
+	containers, err := cli.ContainerList(dockerCtx, types.ContainerListOptions{Filters: filterFeedIn})
 	if err != nil {
 		panic(err)
 	}
@@ -60,20 +59,20 @@ func checkFeederContainers(ctx *cli.Context) {
 	// for each container...
 	for _, container := range containers {
 
-		// cfcLog.Info().Str("container", container.Names[0][1:]).Msg("checking container is running latest feed-in image")
+		cfcLog.Info().Str("container", container.Names[0][1:]).Msg("checking container is running latest feed-in image")
 
-		// // check containers are running latest feed-in image
-		// if container.Image != ctx.String("feedinimage") {
-		cfcLog.Info().Str("container", container.Names[0][1:]).Msg("out of date container being killed for recreation")
-		err := cli.ContainerRemove(dockerCtx, container.ID, types.ContainerRemoveOptions{Force: true})
-		if err != nil {
-			cfcLog.Err(err).Str("container", container.Names[0]).Msg("could not kill out of date container")
-		} else {
-			// avoid killing lots of containers in a short duration
-			cfcLog.Info().Msg("sleep 30 seconds")
-			time.Sleep(30 * time.Second)
+		// check containers are running latest feed-in image
+		if container.Image != ctx.String("feedinimage") {
+			cfcLog.Info().Str("container", container.Names[0][1:]).Msg("out of date container being killed for recreation")
+			err := cli.ContainerRemove(dockerCtx, container.ID, types.ContainerRemoveOptions{Force: true})
+			if err != nil {
+				cfcLog.Err(err).Str("container", container.Names[0]).Msg("could not kill out of date container")
+			} else {
+				// avoid killing lots of containers in a short duration
+				cfcLog.Info().Msg("sleep 30 seconds")
+				time.Sleep(30 * time.Second)
+			}
 		}
-		// }
 	}
 
 	// re-launch this goroutine in 5 mins
